+++ packages/landing/src/index.ts
@@ -1,7 +1,7 @@
 import './global';
 
 import { Basemaps } from './map';
-import { addDebugLayer } from './debug';
+// import { addDebugLayer } from './debug';
 import { WindowUrl } from './url';
 import { BasemapsUi } from './ui';
 import { isWebpSupported } from './webp';
@@ -19,5 +19,5 @@ document.addEventListener('DOMContentLoaded', async () => {
     const ui = new BasemapsUi(basemaps);
     window.basemapsUi = ui;
 
-    if (basemaps.config.debug) addDebugLayer(basemaps);
+    // if (basemaps.config.debug) addDebugLayer(basemaps);
 });