+++ packages/landing/src/ui.ts
@@ -1,5 +1,5 @@
 import { Epsg, GoogleTms, Nztm2000QuadTms, TileMatrixSet } from '@basemaps/geo';
-import { GaEvent, gaEvent } from './config';
+import { Config, GaEvent, gaEvent } from './config';
 import { Basemaps } from './map';
 import { MapOptions, MapOptionType, WindowUrl } from './url';
 
@@ -7,6 +7,7 @@ export class BasemapsUi {
     projectionNztm: HTMLElement;
     projectionWm: HTMLElement;
 
+    apiKey: HTMLElement;
     apiXyz: HTMLElement;
     apiWmts: HTMLElement;
 
@@ -19,6 +20,10 @@ export class BasemapsUi {
 
     constructor(basemaps: Basemaps) {
         this.basemaps = basemaps;
+
+        const versionEl = document.getElementById('basemaps-version');
+        if (versionEl) versionEl.innerText = Config.Version;
+
         this.bindMenuButton();
         this.bindProjectionButtons();
         this.bindApiLinks();
@@ -82,14 +87,17 @@ Your Service/App URL:
 
     bindApiLinks(): void {
         const apiXyz = document.getElementById('api-xyz');
+        const apiKey = document.getElementById('api-key');
         const apiWmts = document.getElementById('api-wmts');
-        if (apiXyz == null || apiWmts == null) {
+        if (apiXyz == null || apiWmts == null || apiKey == null) {
             throw new Error('Unable to find api inputs');
         }
+        this.apiKey = apiKey;
         this.apiXyz = apiXyz;
         this.apiWmts = apiWmts;
         this.bindCopyFromInput(apiXyz);
         this.bindCopyFromInput(apiWmts);
+        this.bindCopyFromInput(apiKey);
     }
 
     /** Attach a listener to a button to copy the nearby input element */
@@ -163,7 +171,8 @@ Your Service/App URL:
         }
         const cfg: MapOptions = { ...this.basemaps.config, tileMatrix };
 
-        this.apiXyz.querySelector('input')!.value = WindowUrl.toTileUrl(cfg, MapOptionType.Tile);
+        this.apiXyz.querySelector('input')!.value = WindowUrl.toTileUrl(cfg, MapOptionType.TileRaster);
         this.apiWmts.querySelector('input')!.value = WindowUrl.toTileUrl(cfg, MapOptionType.Wmts);
+        this.apiKey.querySelector('input')!.value = Config.ApiKey;
     }
 }