+++ packages/landing/src/url.ts
@@ -1,4 +1,4 @@
-import { Epsg, GoogleTms, TileMatrixSet, TileMatrixSets } from '@basemaps/geo';
+import { Epsg, GoogleTms, Nztm2000QuadTms, Nztm2000Tms, TileMatrixSet, TileMatrixSets } from '@basemaps/geo';
 import { Config } from './config';
 
 export interface MapLocation {
@@ -10,12 +10,14 @@ export interface MapLocation {
 export interface MapOptions {
     tag: string;
     imageId: string;
+    style: string;
     tileMatrix: TileMatrixSet;
     /** Is the debug layer enabled @default false */
     debug: boolean;
 }
 export const enum MapOptionType {
-    Tile = 'tile',
+    TileRaster = 'raster',
+    TileVector = 'vector',
     TileWmts = 'tile-wmts',
     Wmts = 'wmts',
     Attribution = 'attribution',
@@ -30,7 +32,7 @@ export const WindowUrl = {
      * TODO do we want to follow this
      */
     toHash(loc: MapLocation): string {
-        return `#@${loc.lat},${loc.lon},z${loc.zoom}`;
+        return `#@${loc.lat.toFixed(7)},${loc.lon.toFixed(7)},z${loc.zoom}`;
     },
 
     /**
@@ -68,15 +70,15 @@ export const WindowUrl = {
         const urlParams = new URLSearchParams(search);
         const tag = urlParams.get('v') ?? 'production';
         const imageId = urlParams.get('i') ?? 'aerial';
+        const style = urlParams.get('s') ?? 'topolike';
         const debug = urlParams.get('debug') != null;
 
         const projectionParam = (urlParams.get('p') ?? GoogleTms.identifier).toLowerCase();
         let tileMatrix = TileMatrixSets.All.find((f) => f.identifier.toLowerCase() === projectionParam);
-        if (tileMatrix == null) {
-            tileMatrix = TileMatrixSets.get(Epsg.parse(projectionParam) ?? Epsg.Google);
-        }
+        if (tileMatrix == null) tileMatrix = TileMatrixSets.get(Epsg.parse(projectionParam) ?? Epsg.Google);
+        if (tileMatrix.identifier === Nztm2000Tms.identifier) tileMatrix = Nztm2000QuadTms;
 
-        return { tag, imageId, tileMatrix, debug };
+        return { tag, imageId, style, tileMatrix, debug };
     },
 
     baseUrl(): string {
@@ -95,8 +97,12 @@ export const WindowUrl = {
         const projectionPath = isDefaultTileMatrix ? tileMatrix.projection.toEpsgString() : tileMatrix.identifier;
         const baseTileUrl = `${this.baseUrl()}/v1/tiles/${opts.imageId}${tag}/${projectionPath}`;
 
-        if (urlType === MapOptionType.Tile) {
+        if (urlType === MapOptionType.TileRaster) {
             return `${baseTileUrl}/{z}/{x}/{y}.${WindowUrl.ImageFormat}${api}`;
+        }
+
+        if (urlType === MapOptionType.TileVector) {
+            return `${baseTileUrl}/style/${opts.style}.json${api}`;
         }
 
         if (urlType === MapOptionType.TileWmts) {